% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{create_plot_dir}
\alias{create_plot_dir}
\title{Creates plot folder in your working directory}
\usage{
create_plot_dir(x)
}
\arguments{
\item{folder}{}
}
\description{
Creates plot folder in your working directory
}
