% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum annotator.R
\name{fragment_map}
\alias{fragment_map}
\title{Fragment Map Plotting function}
\usage{
fragment_map(data = data_merged, subset = FALSE, uvhcpd = FALSE,
  Plot_title_suffix = "", Plot_dir_name = "plot/",
  Plot_file_name = "pin1_6al7_2phos_sty", Plot_width = 3,
  Plot_height = 10)
}
\arguments{
\item{Plot_height}{}
}
\description{
Fragment Map Plotting function
}
