% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{import_masslists}
\alias{import_masslists}
\title{Import masslists output files from TopDownLab}
\usage{
import_masslists(dir = "", folder = "/", bound = FALSE, mod.number = 0,
  shiny = FALSE, files = inFiles)
}
\arguments{
\item{dir}{directory to get files if not specified working directory}

\item{folder}{folder in which files for import are}

\item{bound}{whether output data_list will have files combined e.i. C_0 and N_0}

\item{mod.number}{number of modifications allowed in top-down lab searches and outputs}
}
\description{
Import masslists output files from TopDownLab
}
