% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{prepare_masslists}
\alias{prepare_masslists}
\title{Import masslists and arrange them into the list}
\usage{
prepare_masslists(dir = "", shiny = FALSE)
}
\arguments{
\item{dir}{}
}
\description{
Import masslists and arrange them into the list
}
