% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgmntstats.R
\name{termini_sequence_coverage}
\alias{termini_sequence_coverage}
\title{Terminal Sequence coverages}
\usage{
termini_sequence_coverage(data = data_merged, ppm = 3.5)
}
\arguments{
\item{data}{}
}
\description{
Terminal Sequence coverages
}
