% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum annotator.R
\name{annotate_spectrum}
\alias{annotate_spectrum}
\title{Spectrum annotation}
\usage{
annotate_spectrum(data = gList, dir2 = "", dir3 = "", subset = FALSE,
  mergefrags = FALSE, Plot_file_name = "annotated spectrum",
  ranges = ranges)
}
\arguments{
\item{ranges}{}
}
\description{
Spectrum annotation
}
